remove old StorageManager module
